# internal/infrastructure/http/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateLocationDTO:
    properties:
      address:
        maxLength: 200
        minLength: 5
        type: string
      department:
        type: string
      district:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      postal_code:
        type: string
      province:
        type: string
      public_code:
        maxLength: 50
        minLength: 2
        type: string
      tenant_id:
        type: string
      type:
        enum:
        - FISICO
        - ONLINE
        - AMBOS
        type: string
      url:
        type: string
    required:
    - address
    - department
    - district
    - name
    - province
    - public_code
    - tenant_id
    - type
    type: object
...
  model.EstablishmentType:
    enum:
    - FISICO
    - ONLINE
    - AMBOS
    type: string
    x-enum-varnames:
    - EstablishmentPhysical
    - EstablishmentOnline
    - EstablishmentBoth
  model.Location:
    properties:
      address:
        type: string
      createdAt:
        type: string
      department:
        type: string
      district:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      postalCode:
        type: string
      province:
        type: string
      publicCode:
        type: string
      tenantID:
        type: string
      type:
        $ref: '#/definitions/model.EstablishmentType'
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API para gestión de reclamos multi-tenant
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Book Claims API
  version: "1.0"
paths:
...
  /location/{id}:
    get:
      description: Retrieve location by UUID
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get location by ID
      tags:
      - location
...
  /tenant/{id}/location:
    post:
      consumes:
      - application/json
      description: Create location under a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: 9Location data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLocationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Location'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a location for a tenant
      tags:
      - location
  /tenant/{id}/locations:
    get:
      description: Retrieve all locations under a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Location'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get locations by tenant ID
      tags:
      - location
...
swagger: "2.0"